fix(train/eval): cast bf16 logits to float32 before numpy in metric path
torch.amp.autocast(bfloat16) makes logits bf16; NumPy has no bf16 dtype, so
torch.sigmoid(logits).detach().cpu().numpy() raised 'Got unsupported ScalarType
BFloat16' on the first A100 training step. Add .float() before .numpy() at all
metric/eval conversion sites (train loop, validation loop, evaluate._predict_batch
which tampering_type_analysis reuses, and failure_analysis._predict_failure_batch).
Labels also cast defensively (already float32). Loss/backward/AMP config untouched
(all sites are post-detach / inference_mode, metric-only).

+++ src/evaluate.py
@@ -147,7 +147,7 @@ def _predict_batch(model, batch: list[dict[str, Any]], config: dict[str, Any], r
         with torch.inference_mode():
             with torch.amp.autocast("cuda", dtype=amp_dtype, enabled=runtime["device"] == "cuda"):
                 logits = model(x)
-            probs = torch.sigmoid(logits).detach().cpu().numpy()[:, 0]
+            probs = torch.sigmoid(logits).float().detach().cpu().numpy()[:, 0]
         return [probs[i] for i in range(probs.shape[0])]
     except Exception as exc:
         raise RuntimeError(f"Model inference failed: {exc}") from exc
